requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
mysqlclient==2.1.1        # or pymysql if you prefer: pymysql==1.0.4
gunicorn==20.1.0          # for production/run
